gettext: fix dep on libiconv w/musl
AFAICT this should "always" be provided (gettext dep on libiconv)
but the conditions present are used to avoid a bootstrapping problem
on Linux/glibc since there `iconv == stdenv.cc.libc`.
(see issues linked in expression above modified line)

Fix this on musl by narrowing the check to glibc.

I'm not sure what the Cygwin situation is about, so leaving it alone.