singularity: patch up many more places, likely too many, let's see
"Too many" I mean that only a few of these are host-PATH's,
and so should be patched, with the rest and expected majority
being guest-PATH's (forgiving these terms)-- and should
be left as-is. Where possible defaulting to relevants standards
but given the diversity (if minor) I'm not quite sure without
going through all of this much more closely.

Anyway, mostly I want to see that this fixes the failure to find
mksquashfs in the random example I'm trying, and then will
do something about cleaning this up.  Hopefully. :)