arion-compose: Enable without tests
This adds me as a maintainer of arion-compose, which provides the
executable for the arion package (#71092).

Note that it has a different name because it was already taken on
Hackage before arion switched to Haskell.

(cherry picked from commit 3ca15e5fe10968297ecd9ebabf6c77f3e5f1b3d0)