ell,iwd: don't use clang, badness re __start/__stop
For some reason the start/stop end up also containing a null entry
(entire 24byte struct of 0's) which is not expected by ell/iwd.
Not sure why and feature is barely documented.  Dunno.

First guess would be different flags are given to the linker,
or some other difference re:wrappers and hardening flags;
by first guess I mean it'd be good to check that for sanity :).