lxqt: restrict to linux platform
libqtxdg fails to build on darwin.

LXQt is a DE for linux so assume darwin is not supported.

The platform for some packages that may be useful outside of
LXQt (like pcmanfm-qt, lximage-qt and others) and do not depend on
libqtxdg are not being changed.

(cherry picked from commit a61bd1e42d5512f0e5180edf7ea6090afcd9544b)