tests/systemd-networkd: add regression test for service group
Previously systemd-networkd.service ran as systemd-network:nogroup.
The wireguard private key file is now owned by root:systemd-network with
mode 0640. It is therefore required that the systemd-network user is in the group
with the same name, so that it is able to read the key file.

(cherry picked from commit ec16f4fd5b78edcd11f68cb96b41b9a810af315e)