texlive: bump for latest ctan, texlive.info archive
For some context, this was previously and recently discussed:

https://github.com/NixOS/nixpkgs/pull/61469#issuecomment-549477615
https://github.com/NixOS/nixpkgs/pull/71419#issuecomment-545867115

Fortunately it appears, due to developments since those discussion,
that there is now a versioned mirror for ctan contents!!

Naturally this commit is about moving to that mirror:
updating the package lists and regenerating the hashes.

This has two huge advantages:

* Updating the pinned ctan packages can be done more frequently
  than the current "once a year".
* We don't have to rely on community members to create a snapshot
  and mirror it indefinitely and at their own cost.

Minor "advantage":

* I should be writing but instead I was working on this :P

It appears new, I stumbled upon a mention of it here[1]:

> Native TL updates ultimately come from a directory named tlnet (on CTAN).
> An archive for tlnet is available, with daily updates since
> September 2019, and the final yearly release going back about ten years.

[1] https://www.tug.org/texlive/pkginstall.html

----

Concerns and theirs status:

* Package integrity:
There may be concerns about ensuring the package hashes are obtained
from a trusted source-- I haven't looked into this at all "yet".

* Mirror reliability and not-abusing-nice-things-from-good-people
If the mirror is a small operation it may be hostile to point
the network request fury of the Nixpkgs texlive community ;).
This doesn't appear to be the case and there's no mention of it,
and presumably wouldn't be linked from tug.org if wasn't to be used?
Would be good to chase down an announcement or otherwise ensure
this is expected usage.
Time will tell re:mirror reliability (or speed?),
so far it seems like a solid service meant to be used
and unlikely to be problematic.

* Is the archive only updated once a day?
The archive is organized by date, which is wonderful (!)
but might mean the contents may see changes during the day.
I suspect this is not the case but just a thought.
Regardless, any problems seem easily avoided by simply
not updating to the current day and use yesterday instead.

----

If this is the way we move forward,
there are changes/fixes/improvements to be made
in terms of documentation as well as (possibly)
putting together some update automation tooling.

---

I have only tested this with a single project so far,
and it doesn't have any especially esoteric or demanding
needs.  Help testing by checking your projects to see
that they build and produce expected results.
Building against the before/after set and using `diffpdf`
is how I usually vet these updates, if that's useful.
The `diffpdf` especially helps identify the minor
appearance changes I usually miss visually-- tweaks
to microtype behavior or slightly different layout/placement
of figures or handling of beamer slides, etc.