buildRustCrate: document and cleanup the symbol seeding
That code had been in the derivation for a while but no explanation was
given why that is needed. It might be helpful to our future selfs to
document why things are done the way they are.

(cherry picked from commit f4aeabd04a6f5c0daffebce808e0ea0301266af0)