remmina: 1.3.4 -> 1.3.7
According to the release notes this is a
"This release is meant to prepare the stable release for January, before the freeze in Debian an Ubuntu ".
Not sure what that means, but I expect a stable release
in NixOS 20.03 for sure.

https://gitlab.com/Remmina/Remmina/-/tags/v1.3.7
(cherry picked from commit 586df03d5bda0007a27eb8ef6b04ac1d5802a47c)