flyway: fix jarDirs option
this is not a system property, but rather an argument to
org.flaywaydb.commandline.Main.main().  therefore, it must come after,
rather than before, the name of the main class.  (otherwise it's
interpreted as an argument to the VM itself.)

flyway also expects the option and value to be separated by '=', not ' '.

follows on to #76094

(cherry picked from commit 2867d1963aac8a5f587930644070b7d07e526db3)