python3Packages.srsly: 0.2.0 -> 1.0.1
srsly 2.0.0 is also released. But since spaCy/textacy are the only
dependencies and spaCy requires < 1.1.0, it probably makes more sense
to bump this to 1.0.1 for the time being.

And additional argument in favor of 1.0.1 is that 2.0.0 drops support
for Python 3.5.

(cherry picked from commit 7269a92820ec60967b0475883eb14c00cc118012)