expect: remove hack for aarch64 build
Not sure if I missed something or the issue got fixed later, but it's
not needed anymore to pass a `--build` flag to the configure script on
aarch64.

The autoreconfHooks are still needed for expect and bzip2.

(cherry picked from commit 9646fafc5d182d01288b179115aa43c3d9c2931c)