kitty: remove unused nativeBuildInputs on macOS
These two dependencies were used to generate the PNG icons that would later be converted to the icon for the macOS app. The PNGs are now stored in git and are not generated during the build anymore.

(cherry picked from commit 1b74dfcf6dc16bde182ba80867ec133fa501f60d)