rainicorn: remove broken package that has been abandoned upstream
The upstream project is dead, with the last commit being in 2016:
https://github.com/RustDT/Rainicorn

This no longer builds in NixPkgs, and has been marked as broken for a very long
time. Since the package is not maintained and is dead upstream, let's delete.

(cherry picked from commit cd8c987fbc72beee00cfa0458f272eecfe6991ac)