gcc9: add patches required for target musl
These patches are taken from:
  https://github.com/richfelker/musl-cross-make/tree/master/patches/gcc-9.2.0
and based on a recommendation from the musl maintainers as these patches
are deemed critical for gcc9/musl to work correctly.

This started with an obscure bug I had while investigating the following
program which was getting unexpectedly aborted instead of catching the
exception:
```
\#include <boost/thread.hpp>
\#include <iostream>
\#include <thread>

struct Foo {};

void foo() {
  try {
    std::cout << "entered thread" << std::endl;
    throw Foo();
  } catch (...) {
  }
}

void bar() {
  try {
    foo();
  } catch (...) {
  }
}

int main() {
  std::thread _bar(bar);
  boost::thread _foo(foo);
  _bar.join();
  _foo.join();
  std::cout << "should ignore exception" << std::endl;
  return 0;
}
```

Thanks to nsz and dalias in #musl freenode channel they traced down the
issue to be linked to a missing patch (0014-fix-gthr-weak-refs-for-libgcc.patch
in this case) which was needed. They also recommended adding the following
patches that I added as those are all critical.
I think these patches aren't really musl specific and could be applied
all the time, in fact 0014-fix-gthr-weak-refs-for-libgcc.patch is
already upstream in GCC 10 but I'm not sure if we should do that.